+++ modules/moddevel_txdata.c
@@ -530,9 +530,5 @@ ib_status_t ib_moddevel_txdata_fini(
     ib_engine_t                 *ib,
     ib_module_t                 *mod)
 {
-    ib_hook_tx_unregister(ib,
-                          request_header_finished_event,
-                          tx_header_finished);
-
     return IB_OK;
 }