+++ CHANGES.md
@@ -69,6 +69,8 @@ IronBee v0.8.0
   a single bool; added accessors; `ib_tfn_transform()` has been renamed to `ib_tfn_execute()` and now handles lists properly; separated creation and
   registration similar to operators.
 
+* All ib_hook_xxx_unregister functions have been removed.
+
 **Util**
 
 * Add external iterator support for hash.  See `ib_hash_iterator*`.